:metal: replaced directory name
