+++ Sources/KuriCore/Command/Error/KuriErrorType.swift
