+++ Sources/KuriCore/Command/Generator.swift
@@ -14,22 +14,6 @@ public struct Generator {
     public let argument: GenerateArgument
     public let yamlReader: YamlReader
     
-    private var dateComponent: DateComponent = { _ -> DateComponent in
-        let component = Calendar(identifier: .gregorian).dateComponents([.year, .month, .day], from: Date())
-        guard
-            let year = component.year,
-            let month = component.month,
-            let day = component.day
-            else {
-                fatalError("Can't get system date")
-        }
-        return DateComponent(year: year, month: month, day: day)
-    }(())
-    
-    private var userName: String {
-        return run(bash: "echo $USER").stdout
-    }
-    
     public init(
         argument: GenerateArgument,
         yamlReader: YamlReader
@@ -171,28 +155,6 @@ extension Generator {
 }
 
 fileprivate extension Generator {
-    struct DateComponent {
-        let year: Int
-        let month: Int
-        let day: Int
-        
-        var date: String {
-            return "\(year)/\(month)/\(day)"
-        }
-    }
-    
-    private func replaceEnvironmentText(content: String, prefix: String, targetName: String, userName: String, date: DateComponent) -> String {
-        let replacedContent = content
-            .replacingOccurrences(of: "__PREFIX__", with: prefix)
-            .replacingOccurrences(of: "__TARGET__", with: targetName)
-            .replacingOccurrences(of: "__USERNAME__", with: userName)
-            .replacingOccurrences(of: "__DATE__", with: date.date)
-            .replacingOccurrences(of: "__YEAR__", with: "\(date.year)")
-            .replacingOccurrences(of: "__MONTH__", with: "\(date.month)")
-            .replacingOccurrences(of: "__DAY__", with: "\(date.day)")
-        return replacedContent
-    }
-
     fileprivate func generate(with prefix: String, for components: [GenerateComponent], and templateHeadPath: String) throws {
         print("Begin generate")
         defer {
@@ -207,8 +169,9 @@ fileprivate extension Generator {
             let projectRootPath = yamlReader.projectRootPath(for: typeFor)
             let projectFileName = yamlReader.projectFileName(for: typeFor)
             
+            let targetName = yamlReader.targetName(for: typeFor)
             let projectFilePath = projectRootPath + projectFileName + "/"
-            let directoryPath = generateRootPath + component.generateDirectoryPath.joined(separator: "/") + "/"
+            let directoryPath = generateRootPath + component.makeDirectoryPath(prefix: prefix, targetName: targetName).joined(separator: "/") + "/"
             let filePath = directoryPath + prefix + component.templateFileName
             
             let project: XcodeProject
@@ -226,8 +189,7 @@ fileprivate extension Generator {
                 print("can't find: \(componentType)")
                 return
             }
-            let targetName = yamlReader.targetName(for: typeFor)
-            let writeCotent = replaceEnvironmentText(content: templateContent, prefix: prefix, targetName: targetName, userName: userName, date: dateComponent)
+            let writeCotent = templateContent.replaceEnvironmentText(prefix: prefix, targetName: targetName)
             try fileOperator.createDirectory(for: directoryPath)
             fileOperator.createFile(for: filePath)
