+++ Sources/KuriCore/Command/SetupComponentType.swift
