+++ Sources/KuriCore/Util/Helper.swift
